canvis grupals Tools, Writer i Reader
