+++ src/main/java/org/develop/Decoration.java
@@ -11,8 +11,8 @@ public class Decoration extends Product {
     }
 
     //Constructor pel mètode JSONArrayToHashMap
-    public Decoration(String ref, String name, int quantity, double price, MaterialType material) {
-        super(ref, name, quantity, price, ProductType.DECORATION);
+    public Decoration(int ID, String ref, String name, int quantity, double price, MaterialType material) {
+        super(ID, ref, name, quantity, price, ProductType.DECORATION);
         this.material = material;
     }
