+++ src/main/java/org/develop/Product.java
@@ -20,8 +20,8 @@ public class Product {
     }
 
     //Constructor pel mètode JSONArrayToHashMap
-    public Product(String ref, String name, int quantity, double price, ProductType productType) {
-        this.ID = ++idCounter;
+    public Product(int ID, String ref, String name, int quantity, double price, ProductType productType) {
+        this.ID = ID;
         this.ref = ref;
         this.name = name;
         this.quantity = quantity;
